fix(wiki): one identity colour for a peer's caret and their presence chip
Confirmed: a peer's caret and their presence-chip ring were derived two
different ways, so they disagreed.

  caret (presence.ts)          PEER_COLORS — 8 raw hex   keyed by hash(userId)
  chip ring (PresenceAvatars)  USER_COLORS — 5 Opal vars keyed by list index

Different palettes and different keys, so agreement was coincidence — a hue
that happened to look close, not a match.

Keying the ring on list position is a second defect on its own. The list has
the current user filtered out and is in participant order, so a peer's ring
colour changed when somebody *else* joined or left ahead of them: with a
four-person roster, dropping the first participant moved one peer's ring from
lime to yellow while that peer did nothing. A colour that identifies someone
cannot move when the roster around them moves.

Both now come from `colorFor` in one module, hashed on `userId`. The palette is
the chip's existing Opal tokens rather than the caret's hex literals — one list
to change, and no raw colours (CLAUDE.md). That drops the caret palette from 8
entries to 5, so distinct users collide a little sooner; a brief ambiguity
between two peers costs less than a colour the reader cannot rely on at all.
